Move `ScheduleController` to ContentModerator controllers
